+++ mountwizzard/support/model_thread.py
@@ -38,7 +38,7 @@ class Model(QtCore.QThread):
     signalModelRedraw = QtCore.pyqtSignal(bool, name='ModelRedrawPoints')
 
     BLUE = 'background-color: rgb(42, 130, 218)'
-    RED = 'background-color: red'
+    RED = 'background-color: red;'
     DEFAULT = 'background-color: rgb(32,32,32); color: rgb(192,192,192)'
     REF_PICTURE = '/model001.py'